Doc string updates and refernce fixes
