+++ boa/instantiation_base.py
@@ -4,7 +4,7 @@ from ax import Objective
 from ax.core.objective import ScalarizedObjective
 from ax.service.utils.instantiation import InstantiationBase
 
-from boa.metrics.modular_metric import get_metric_from_config
+from boa.metrics.metrics import get_metric_from_config
 
 
 class BoaInstantiationBase(InstantiationBase):
