+++ boa/metrics/synthetic_funcs.py
@@ -4,10 +4,18 @@ Synthetic Function
 ########################
 
 """
+from __future__ import annotations
 
+import sys
+from inspect import isclass
+
+import ax.utils
+import botorch.test_functions
 from ax.utils.measurement.synthetic_functions import from_botorch
 from botorch.test_functions.synthetic import Hartmann
 from torch import Tensor
+
+from boa.metrics.modular_metric import ModularMetric
 
 
 class Hartmann4(Hartmann):
@@ -28,3 +36,33 @@ class Hartmann4(Hartmann):
 
 
 hartmann4 = from_botorch(Hartmann4())
+
+
+def get_synth_func(synthetic_metric: str) -> type:
+    synthetic_funcs_modules = [
+        sys.modules[__name__],  # this module
+        ax.utils.measurement.synthetic_functions,
+        botorch.test_functions.synthetic,
+    ]
+    for module in synthetic_funcs_modules:
+        try:
+            return getattr(module, synthetic_metric)
+        except AttributeError:
+            continue
+    # If we don't find the class by the end of the modules, raise attribute error
+    raise AttributeError(f"boa synthetic function: {synthetic_metric} not found in modules: {synthetic_funcs_modules}!")
+
+
+def setup_synthetic_metric(synthetic_metric, instantiate=True, **kw):
+    metric = get_synth_func(synthetic_metric)
+
+    if isclass(metric) and issubclass(metric, ax.utils.measurement.synthetic_functions.SyntheticFunction):
+        metric = metric()  # if they pass a ax synthetic metric class, not instance
+    elif isclass(metric) and issubclass(metric, botorch.test_functions.synthetic.SyntheticTestFunction):
+        # botorch synthetic functions need to be converted
+        metric = from_botorch(botorch_synthetic_function=metric())
+
+    def modular_synthetic_metric(**kwargs):
+        return ModularMetric(**{**kw, **kwargs, "metric_to_eval": metric})
+
+    return modular_synthetic_metric(**kw) if instantiate else modular_synthetic_metric
