+++ docs/user_guide/run_script.rst
@@ -1,5 +1,3 @@
-.. _run_script:
-
 #############
 Run Script
 #############
